# openapi.yaml
openapi: 3.0.0
info:
    title: AirClass API
    description: API for managing classrooms, attendance, and educational resources
    version: 1.0.0
    contact:
        name: AirClass Team
        email: support@airclass.com
servers:
    - url: /airclass-api
      description: API base URL
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
        ApiResponse:
            type: object
            required:
                - status
                - message
            properties:
                status:
                    type: boolean
                    description: Indicates if the request was successful
                message:
                    type: string
                    description: Response message
                data:
                    type: object
                    description: Response data (if any)
                    nullable: true
...
        Image:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                classroom_id:
                    type: integer
                    format: int64
                full_path:
                    type: string
                created_at:
                    type: string
                    format: date-time
...
paths:
...
    /image:
        get:
            summary: List images
            tags: [Image]
            security:
                - BearerAuth: []
            parameters:
                - name: classroom_id
                  in: query
                  schema:
                      type: integer
                  description: Filter by classroom ID
                - name: id
                  in: query
                  schema:
                      type: integer
                  description: Filter by image ID
            responses:
                "200":
                    description: List of images
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/ApiResponse"
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: "#/components/schemas/Image"
        post:
            summary: Upload image
            tags: [Image]
            security:
                - BearerAuth: []
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required:
                                - classroom_id
                                - image
                            properties:
                                classroom_id:
                                    type: integer
                                image:
                                    type: string
                                    format: binary
            responses:
                "201":
                    description: Image uploaded
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/ApiResponse"
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  image_id:
                                                      type: integer
                                                  path:
                                                      type: string
...